refact: rename Column components to be generic
