+++ src/components/HomePartner/index.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from "react"
 import * as SC from "./styles"
 
-export const Partner: FC = () => {
+export const HomePartner: FC = () => {
   return (
-    <SC.PartnerWrapper>
+    <SC.HomePartnerWrapper>
       Member of
       <SC.StyledJetBrainsLogo alt="JetBrains" />
       Supported user groups
-    </SC.PartnerWrapper>
+    </SC.HomePartnerWrapper>
   )
 }
