+++ src/components/HomePartner/styles.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components"
 import JetBrainsLogo from "../../images/jetbrains-logo.svg"
 
-export const PartnerWrapper = styled.div`
+export const HomePartnerWrapper = styled.div`
   margin-top: 32px;
   display: flex;
   align-items: center;
