+++ src/templates/languagePost.tsx
@@ -1,6 +1,6 @@
 import { graphql } from "gatsby"
 import React, { FC, Fragment } from "react"
-import { ColumnHeader } from "../components/ColumnHeader"
+import { Header } from "../components/Header"
 import { Markdown } from "../components/Markdown"
 import { SEO } from "../components/SEO"
 
@@ -14,7 +14,7 @@ const LanguagePost: FC<any> = ({ data }) => {
   return (
     <Fragment>
       <SEO title={frontmatter.title} description={excerpt} />
-      <ColumnHeader
+      <Header
         relativePath={relativePath}
         basePath="/resources"
         title={frontmatter.title}
